add enc-dec model tests
